Refactor: Rename `AbstractEmitter` to `LanguageEmitter` for better clarity and consistency
