+++ src/compiler/core/src/commonMain/kotlin/community/flock/wirespec/compiler/core/emit/LanguageEmitter.kt
@@ -12,7 +12,7 @@ import community.flock.wirespec.compiler.core.parse.Type
 import community.flock.wirespec.compiler.core.parse.Union
 import community.flock.wirespec.compiler.utils.Logger
 
-abstract class AbstractEmitter :
+abstract class LanguageEmitter :
     Emitter,
     Emitters,
     BaseEmitter,
