+++ src/compiler/emitters/java/src/commonMain/kotlin/community/flock/wirespec/emitters/java/JavaEndpointDefinitionEmitter.kt
@@ -1,9 +1,9 @@
 package community.flock.wirespec.emitters.java
 
 import community.flock.wirespec.compiler.core.concatGenerics
-import community.flock.wirespec.compiler.core.emit.AbstractEmitter.Companion.firstToLower
-import community.flock.wirespec.compiler.core.emit.AbstractEmitter.Companion.firstToUpper
-import community.flock.wirespec.compiler.core.emit.AbstractEmitter.Companion.isStatusCode
+import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.firstToLower
+import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.firstToUpper
+import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.isStatusCode
 import community.flock.wirespec.compiler.core.emit.EndpointDefinitionEmitter
 import community.flock.wirespec.compiler.core.emit.EndpointEmitter
 import community.flock.wirespec.compiler.core.emit.IdentifierEmitter
