+++ src/compiler/emitters/typescript/src/commonMain/kotlin/community/flock/wirespec/emitters/typescript/TypeScriptEndpointDefinitionEmitter.kt
@@ -1,7 +1,7 @@
 package community.flock.wirespec.emitters.typescript
 
-import community.flock.wirespec.compiler.core.emit.AbstractEmitter.Companion.firstToLower
-import community.flock.wirespec.compiler.core.emit.AbstractEmitter.Companion.firstToUpper
+import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.firstToLower
+import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.firstToUpper
 import community.flock.wirespec.compiler.core.emit.EndpointDefinitionEmitter
 import community.flock.wirespec.compiler.core.emit.EndpointEmitter
 import community.flock.wirespec.compiler.core.emit.ImportEmitter
