+++ src/converter/openapi/src/commonMain/kotlin/community/flock/wirespec/openapi/v2/OpenAPIV2Emitter.kt
@@ -16,7 +16,7 @@ import community.flock.kotlinx.openapi.bindings.v2.SchemaObject
 import community.flock.kotlinx.openapi.bindings.v2.SchemaOrReferenceObject
 import community.flock.kotlinx.openapi.bindings.v2.StatusCode
 import community.flock.kotlinx.openapi.bindings.v2.SwaggerObject
-import community.flock.wirespec.compiler.core.emit.AbstractEmitter
+import community.flock.wirespec.compiler.core.emit.LanguageEmitter
 import community.flock.wirespec.compiler.core.emit.Emitted
 import community.flock.wirespec.compiler.core.emit.FileExtension
 import community.flock.wirespec.compiler.core.parse.Channel
@@ -35,7 +35,7 @@ import kotlinx.serialization.encodeToString
 import kotlinx.serialization.json.JsonPrimitive
 import community.flock.kotlinx.openapi.bindings.v2.Type as OpenAPIType
 
-object OpenAPIV2Emitter : AbstractEmitter() {
+object OpenAPIV2Emitter : LanguageEmitter() {
 
     override val extension = FileExtension.JSON
