+++ src/converter/openapi/src/commonMain/kotlin/community/flock/wirespec/openapi/v2/OpenAPIV2Parser.kt
@@ -23,7 +23,7 @@ import community.flock.kotlinx.openapi.bindings.v2.SchemaOrReferenceOrBooleanObj
 import community.flock.kotlinx.openapi.bindings.v2.StatusCode
 import community.flock.kotlinx.openapi.bindings.v2.SwaggerObject
 import community.flock.wirespec.compiler.core.ModuleContent
-import community.flock.wirespec.compiler.core.emit.AbstractEmitter.Companion.firstToUpper
+import community.flock.wirespec.compiler.core.emit.LanguageEmitter.Companion.firstToUpper
 import community.flock.wirespec.compiler.core.parse.AST
 import community.flock.wirespec.compiler.core.parse.Definition
 import community.flock.wirespec.compiler.core.parse.DefinitionIdentifier
